% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateFeatures.R
\name{aggregateFeatures}
\alias{aggregateFeatures}
\alias{aggregateFeatures,Features-method}
\title{Aggregate an assay's quantitative features}
\usage{
\S4method{aggregateFeatures}{Features}(object, i, fcol, name = "newAssay", fun = median, ...)
}
\arguments{
\item{object}{An instance of class \link{Features}.}

\item{i}{The index or name of the assay which features will be
aggregated the create the new assay.}

\item{fcol}{The feature variable of assay \code{i} defining how to
summerise the features.}

\item{name}{A \code{character(1)} naming the new assay. Default is
\code{newAssay}. Note that the function will fail if there's
already an assay with \code{name}.}

\item{fun}{A function used for quantitative feature
aggregation. Default is \code{median}.}

\item{...}{Additional parameters passed the \code{fun}.}
}
\value{
A \code{Features} object with an additional assay.
}
\description{
This function aggregates the quantitative features of an assay,
applying a summarisation function (\code{fun}) to sets of features as
defined by the \code{fcol} feature variable. The new assay's features
will be named based on the unique \code{fcol} values.
}
\examples{

## An example Features with PSM-level data
data(feat1)

## Aggregate PSMs into peptides
feat1 <- aggregateFeatures(feat1, "psms", "Sequence", name = "peptides")
feat1

## Aggregate peptides into proteins
feat1 <- aggregateFeatures(feat1, "peptides", "Protein", name = "proteins")
feat1
}
\seealso{
The \code{Features} vignette provides an extended example.
}
