% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{featureData}
\alias{featureNames}
\alias{featureVariables}
\alias{features}
\alias{tidyFeatureData}
\title{Placeholder for generics functions documentation}
\usage{
featureData(x, i, ...)

featureNames(x, i, ...)

featureVariables(x, i, ...)

features(x, i, ...)

tidyFeatureData(object, i, ...)
}
\arguments{
\item{x}{A parameter.}

\item{i}{Another parameter.}

\item{...}{Additional parameters.}
}
\description{
Placeholder for generics functions documentation
}
