% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feat3}
\alias{feat3}
\title{Example \code{QFeatures} object after processing}
\format{
An object of class \code{QFeatures} of length 7.
}
\source{
The \code{pRolocdata} package: \url{http://bioconductor.org/packages/pRolocdata/}
}
\usage{
feat3
}
\description{
\code{feat3} is a small \code{QFeatures} object  dataset can be is an example that tests QFeatures as it could
routinely be used within a pipeline. Note that the 3 special cases
of AssayLinks are present in this dataset:
\itemize{
\item One to one link: link between 1 parent and 1 child with one to
one row mapping. E.g. "peptides" to "normpeptides"
\item One parent to multiple children: link between 1 parents and
multiple children mapping. E.g. "peptides" to "normpeptides" and
"proteins"
\item Multiple parents to one child: link between multiple parents and
one child mapping. E.g. "psms1" and "psms2" to "psmsall"
}
}
\details{
Run \code{plot(feat3)} for an overview of the relationships
}
\examples{

data("feat3")
plot(feat3)

}
\seealso{
See \link{feat1} for other example and test data sets.
}
\keyword{datasets}
