% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feat4}
\alias{feat4}
\title{Example \code{QFeatures}}
\format{
An object of class \code{QFeatures} of length 2.
}
\source{
\code{feat4} was built from \code{feat3}. The source code is available in
\href{https://github.com/rformassspectrometry/QFeatures/blob/master/inst/scripts/make-feat4.R}{\code{inst/scripts/make-feat4.R}}
}
\usage{
feat4
}
\description{
\code{feat3} is a small \code{QFeatures} object that contains 2 PSM-level assays used
to illustrate to creation of unique precursor identifiers and merging, as
shown in \code{\link[=createPrecursorId]{createPrecursorId()}}.
}
\examples{

data("feat4")
feat4

}
\seealso{
See \code{?feat1} for other example/test data sets.
}
\keyword{datasets}
