% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureList-class.R
\docType{class}
\name{FeatureList}
\alias{FeatureList}
\alias{FeatureList-class}
\alias{class:FeatureList}
\title{Contains a list of \code{FetureSet} instances}
\usage{
FeatureList(...)
}
\arguments{
\item{...}{Individual `FeatureSet` instances or a list thereof.}
}
\description{
This data structure as for sole purpose to store a list of
\link{FeatureSet} instances. All \code{FeatureSet} instances in a
\code{FeatureList} must however have the same number of columns.
Objects of the class can be constructed with the \code{FeatureList()}
constructor. The \code{[[} and \code{[} operators work as expected.
}
\examples{

## An empty FeatureList
FeatureList()

## See ?Features for more examples
}
\seealso{
\link{Features} is the main data struture to manipulate and
process quantitative data.
}
\author{
Laurent Gatto
}
