% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features-missing-data.R
\name{missing-data}
\alias{missing-data}
\alias{zeroIsNA}
\alias{zeroIsNA,SummarizedExperiment,missing-method}
\alias{zeroIsNA,Features,missing-method}
\alias{zeroIsNA,Features,numeric-method}
\alias{zeroIsNA,Features,integer-method}
\alias{zeroIsNA,Features,character-method}
\alias{nNA}
\alias{nNA,SummarizedExperiment,missing-method}
\alias{nNA,Features,missing-method}
\alias{nNA,Features,numeric-method}
\alias{nNA,Features,integer-method}
\alias{nNA,Features,character-method}
\alias{filterNA}
\alias{filterNA,SummarizedExperiment-method}
\alias{filterNA,Features-method}
\title{Managing missing data}
\usage{
\S4method{zeroIsNA}{SummarizedExperiment,missing}(object, i)

\S4method{zeroIsNA}{Features,missing}(object, i)

\S4method{zeroIsNA}{Features,integer}(object, i)

\S4method{zeroIsNA}{Features,numeric}(object, i)

\S4method{zeroIsNA}{Features,character}(object, i)

\S4method{nNA}{SummarizedExperiment,missing}(object, i)

\S4method{nNA}{Features,integer}(object, i)

\S4method{nNA}{Features,numeric}(object, i)

\S4method{nNA}{Features,character}(object, i)

\S4method{nNA}{Features,missing}(object, i)

\S4method{filterNA}{SummarizedExperiment}(object, pNA = 0)

\S4method{filterNA}{Features}(object, pNA = 0, i)
}
\arguments{
\item{object}{An object of class \code{Features} or \code{SummarizedExperiment}.}

\item{i}{The index or name of the assay to be processed.}

\item{pNA}{\code{numeric(1)} providing the maximim percentage of
missing values per feature (row) that is acceptable. Feature
with higher percentages are removed. If 0 (default), features
that contain any number of \code{NA} values are dropped.}

\item{...}{Additional arguments.}
}
\value{
An instance of the same class as \code{object}.
}
\description{
This manual page describes the handling of missing values in
\link{Features} objects. In the following functions, if \code{object} is of
class \code{Features}, and optional assay index or name \code{i} can be
specified to define the assay (by name of index) on which to
operate.

The following functions are currently available:
\itemize{
\item \code{zeroIsNA(object, i)} replaces all 0 in \code{object} by \code{NA}. This
is often necessary when third-party software assume that
features that weren't quantified should be assigned an
intensity of 0.
\item \code{nNA(object, i)} return a list of missing value summaries. The
first element \code{nNA} gives the percentage of missing values; the
second element \code{nNArows} provides a table of the number of
missing values for the features (rows) of the assay(s); the
third element \code{nNAcols} provides the number of missing values in
each sample of the assay(s).
\item \code{filterNA(object, pNA, i)} removes features (rows) that contain
\code{pNA} percentage or more missing values.
}

See the \emph{Processing} vignette for examples.
}
