% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features-processing.R
\name{Features-processing}
\alias{Features-processing}
\alias{logTransform}
\alias{logTransform,SummarizedExperiment-method}
\alias{logTransform,Features-method}
\title{Features processing}
\usage{
\S4method{logTransform}{SummarizedExperiment}(object, base = 2, pc = 0)

\S4method{logTransform}{Features}(object, base = 2, i, pc = 0)
}
\arguments{
\item{object}{An object of class \code{Features} or \code{SummarizedExperiment}.}

\item{base}{\code{numeric(1)} providing the base with respect to which
logarithms are computed. Defaults is 2.}

\item{pc}{\code{numeric(1)} with a pseudocount to add to the
quantitative data. Useful when (true) 0 are present in the
data. Default is 0 (no effect).}

\item{i}{The index or name of the assay to be processed.}
}
\description{
This manual page describes common quantitative proteomics data
processing methods using \link{Features} objects. In the following
functions, if \code{object} is of class \code{Features}, and optional assay
index or name \code{i} can be specified to define the assay (by name of
index) on which to operate.

The following functions are currently available:
\itemize{
\item \code{logTransform(object, base = 2, i)}
}
}
