% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSet-class.R
\docType{methods}
\name{FeatureSet}
\alias{FeatureSet}
\alias{FeatureSet-class}
\alias{class:FeatureSet}
\alias{sampleNames,FeatureSet-method}
\alias{sampleNames<-,FeatureSet,character-method}
\alias{dim,FeatureSet-method}
\alias{ncol,FeatureSet-method}
\alias{nrow,FeatureSet-method}
\alias{assay,FeatureSet,ANY-method}
\alias{featureVariables,FeatureSet,missing-method}
\alias{featureData,FeatureSet,missing-method}
\alias{featureNames,FeatureSet,missing-method}
\alias{[,FeatureSet,ANY,ANY,missing-method}
\title{Simlple container for an assay and its featre metadata}
\usage{
FeatureSet(assay = matrix(ncol = 0, nrow = 0),
  featureData = DataFrame())

\S4method{sampleNames}{FeatureSet}(object)

\S4method{sampleNames}{FeatureSet,character}(object) <- value

\S4method{dim}{FeatureSet}(x)

\S4method{ncol}{FeatureSet}(x)

\S4method{nrow}{FeatureSet}(x)

\S4method{assay}{FeatureSet,ANY}(x, i, ...)

\S4method{featureVariables}{FeatureSet,missing}(x, i, ...)

\S4method{featureData}{FeatureSet,missing}(x, i, ...)

\S4method{featureData}{FeatureSet,missing}(x, i, ...)

\S4method{featureNames}{FeatureSet,missing}(x, i, ...)

\S4method{[}{FeatureSet,ANY,ANY,missing}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{assay}{A `matrix` containing the quantitation data.}

\item{featureData}{A `DataFrame` with feature annotations.}

\item{object}{A `FeatureSet` object.}

\item{value}{An appropriate replacement value.}

\item{x}{The `FeatureSet` to subset.}

\item{i}{Subsetting vector for the assay and featureData's rows.}

\item{...}{Additional paramaeters (ignored).}

\item{j}{Subsetting vector for the assay's columns.}

\item{drop}{Always set to `FALSE`.}
}
\description{
A \code{FeatureSet} is a simple container to store quantitative data
(called the \code{assay}) and the associated feature metadata (called
the \code{featureData}). The number of rows and the rownames of the
assay and feature metadata must match exactly.
}
\section{Slots}{

\describe{
\item{\code{assay}}{A \code{matrix} of numerics that stores quantitative data
for a set of features (along the rows) and samples (along the
columns).}

\item{\code{featureData}}{A \code{DataFrame} to hold the features annotations.}

\item{\code{id}}{An \code{integer(1)} holding the set's identifier. For
internal use at the \link{Features}-level only.}

\item{\code{from}}{An \code{integer(1)} refering to the identifier from which
the current object originates. For internal use at the
\link{Features}-level only.}

\item{\code{version}}{A \code{character(1)} providing the class version. For
internal use only.}
}}

\examples{

## An empty FeatureSet
FeatureSet()

## See ?Features for more examples
}
\seealso{
\link{Features} is the main data struture to manipulate and
process quantitative data.
}
\author{
Laurent Gatto
}
