% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSet-links.R
\name{get_featureSet_ids}
\alias{get_featureSet_ids}
\alias{set_featureSet_ids}
\alias{get_next_featureSet_id}
\title{Returns the \link{FeatureSet}s \code{id} slots.}
\usage{
get_featureSet_ids(object)

set_featureSet_ids(object)

get_next_featureSet_id(object)
}
\arguments{
\item{object}{A Features object.}
}
\value{
The \link{FeatureSet}s \code{id} slots.

A Features object with FeatureSet element ids set to
\code{1:length(object)}.

The next `id` in the [Features], as an `integer(1)`.
}
\description{
Returns the \link{FeatureSet}s \code{id} slots.

This function sets the \link{FeatureSet}'s \code{id} slot to
\code{1:length(object)}. The \link{FeatureSet} with the larges number of
features gets \code{id} 1, next one 2, ... and the smallest one get
\code{length(object)}.

Returns the next identifier.
}
