% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlpsms}
\alias{hlpsms}
\title{hyperLOPIT PSM-level expression data}
\format{An object of class \code{data.frame} with 32776 rows and 28 columns.}
\source{
The `pRolocdata` package: \url{http://bioconductor.org/packages/pRolocdata/}
}
\usage{
hlpsms
}
\description{
A `data.frame` with PSM-level quantitation data by Christoforou *et al.*
(2016). This is the first replicate of a spatial proteomics dataset from a
hyperLOPIT experimental design on Mouse E14TG2a embryonic stem
cells. Normalised intensities for proteins for TMT 10-plex labelled
fractions are available for 3 replicates acquired in MS3 mode using an
Orbitrap Fusion mass-spectrometer.
}
\details{
For further details, install the `pRolocdata` package and see
`?hyperLOPIT2015`.
}
\references{
*A draft map of the mouse pluripotent stem cell spatial proteome*
Christoforou A, Mulvey CM, Breckels LM, Geladaki A, Hurrell T, Hayward PC,
Naake T, Gatto L, Viner R, Martinez Arias A, Lilley KS. Nat Commun. 2016 Jan
12;7:8992. doi: 10.1038/ncomms9992. PubMed PMID: 26754106; PubMed Central
PMCID: PMC4729960.
}
\seealso{
See [Features] to import this data using the [readFeatures()] function.
}
\keyword{datasets}
