% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/missing-data.R
\name{zeroIsNA}
\alias{zeroIsNA}
\alias{missing-data}
\alias{zeroIsNA,SummarizedExperiment,missing-method}
\alias{zeroIsNA,Features,missing-method}
\alias{zeroIsNA,Features,numeric-method}
\alias{zeroIsNA,Features,integer-method}
\alias{zeroIsNA,Features,character-method}
\title{Managing missing data}
\usage{
zeroIsNA(object, i)
}
\arguments{
\item{object}{An object of class \code{Features} or \code{SummarizedExperiment}.}

\item{i}{The index or name of the assay to be processed.}
}
\value{
An instance of the same class as \code{object}.
}
\description{
This manual page describes the handling of missing values in
\code{Features} objects. The current functions are available:
\itemize{
\item \code{zeroIsNA(object)}: this function replaces all 0 in \code{object} by
\code{NA}. This is often necessary when third-party software assume
that features that weren't quantified should be assigned an
intensity of 0.
}
}
