% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures-class.R
\name{setQFeaturesType}
\alias{setQFeaturesType}
\alias{getQFeaturesType}
\alias{validQFeaturesType}
\title{Set and Get QFeatures Type}
\usage{
setQFeaturesType(object, type)

getQFeaturesType(object)

validQFeaturesType()
}
\arguments{
\item{object}{An instance of class \link{QFeatures}.}

\item{type}{\code{character(1)} defining the type of the QFeatures.
Must be one of the values returned by \code{\link[=validQFeaturesType]{validQFeaturesType()}}.}
}
\value{
\itemize{
\item \code{setQFeaturesType()}: returns the updated \code{QFeatures} object with
the type stored in its metadata.
\item \code{getQFeaturesType()}: returns a character string indicating the
type of the \code{QFeatures} object. If no type is explicitly set,
it is inferred from the class of the experiments.
\item \code{validQFeaturesType()}: character vector of valid QFeatures types.
}
}
\description{
Developer-level functions to set and retrieve the type of a \code{QFeatures}
object. This type can help internal methods adapt their behavior to the
structure of the data.
}
\details{
These functions control an internal metadata field (\code{._type}) used to
distinguish between different structural uses of \code{QFeatures} objects.
}
\note{
These functions are intended for package developers and internal use.
End users should typically not call them directly.
}
\keyword{internal}
