% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features-constructors.R
\name{readFeatures}
\alias{readFeatures}
\title{Features from tabular data}
\usage{
readFeatures(table, ecol, fnames, ..., name = NULL)
}
\arguments{
\item{table}{File or object holding the quantitative data. Can be
either a `character(1)` with the path to a text-based
spreadsheet (comma-separated values by default, but see `...`)
or an object that can be coerced to a `data.frame`.}

\item{ecol}{A `numeric` indicating the indices of the columns to
be used as expression values. Can also be a `character`
indicating the names of the columns. Caution must be taken if
the column names are composed of special characters like `(`
or `-` that will be converted to a `.` by the `read.csv`
function. If `ecol` does not match, the error message will
dislpay the column names as seen by the `read.csv` function.}

\item{fnames}{An optional `character(1)` or `numeric(1)`
indicating the column to be used as feature names.}

\item{...}{Further arguments that can be passed on to `read.csv`.}

\item{name}{An optional `character(1)` to name the [FeatureSet] in
the resulting [Features] object.}
}
\value{
An instance of class [Features].
}
\description{
Convert tabular data from a spreadsheet or a `data.frame` into a
`Features` object.
}
\seealso{
The [Features] class for an example on how to manipulate use
    `readFeatures` and how to further manipulate the resulting data.
}
\author{
Laurent Gatto
}
