% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{tidyFeatureData,FeatureSet,missing-method}
\alias{tidyFeatureData,FeatureSet,missing-method}
\alias{tidyFeatureData,Features,character-method}
\alias{tidyFeatureData}
\alias{tidyFeatureData,Features,numeric-method}
\title{Manually review feature variables}
\usage{
\S4method{tidyFeatureData}{FeatureSet,missing}(object, i, ...)

\S4method{tidyFeatureData}{Features,character}(object, i, ...)

\S4method{tidyFeatureData}{Features,numeric}(object, i, ...)
}
\arguments{
\item{object}{An instance of class \link{FeatureSet} or \link{Features}.}

\item{i}{For \link{Features} objects only, defines which \code{featureData}
to review.}

\item{...}{Ignored.}
}
\value{
An updated instance of the same class as `object.
}
\description{
This function offers a user to either keep, drop or summarise a
feature variable to amend a \link{FeatureSet}'s or a \link{Features}'
\code{featureData} slot. This is necessary when proceeding with
multiple calls to \code{combineFeatures} to avoid overly intricated
'lists of lists' variables.
}
