% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{featureData}
\alias{featureNames}
\alias{featureVariables}
\title{Placeholder for generics functions documentation}
\usage{
featureData(x, i, ...)

featureNames(x, i, ...)

featureVariables(x, i, ...)
}
\description{
Placeholder for generics functions documentation
}
