% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\docType{class}
\name{AssayLinks-class}
\alias{AssayLinks-class}
\title{Links between Assays}
\description{
This class handles the links different assays within a Features
object. It isn't meant to be used directly by users.
}
\section{Slots}{

\describe{
\item{\code{id}}{An \code{integer(1)} holding the set's identifier. For
internal use at the \link{Features}-level only.}

\item{\code{from}}{An \code{integer(1)} refering to the identifier from which
the current object originates. For internal use at the
\link{Features}-level only.}

\item{\code{fcol}}{A \code{character(1)} storing the feature variables that was
used to group the object's parent features.}
}}

