% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayLinks.R
\docType{methods}
\name{AssayLinks}
\alias{AssayLinks}
\alias{AssayLink}
\alias{AssayLink-class}
\alias{AssayLinks-class}
\alias{class:AssayLinks}
\alias{class:AssayLink}
\alias{show,AssayLink-method}
\title{Links between Assays}
\usage{
AssayLink(name, from = NA_character_, fcol = NA_character_,
  hits = Hits())

\S4method{show}{AssayLink}(object)

AssayLinks(..., names = NULL)
}
\arguments{
\item{name}{A mandatory name of the assay(s).}

\item{from}{The name of the parent assay, or \code{NA_character_}, if
not applicable.}

\item{fcol}{The feature variable of the parent assay used to
generate the current assay (used in
\code{combineFeatures}). \code{NA_character_}, if not applicable.}

\item{hits}{An object of class \link[S4Vectors:Hits]{S4Vectors::Hits} matching the
features of two assays.}

\item{object}{An `AssayLink` object to show.}

\item{...}{A set of \code{AssayLink} objects or a list thereof.}

\item{names}{A \code{character()} of \code{AssayLink} names. If provided,
\code{...} are ignored, and \code{names} is used to create an
\code{AssayLinks} object with \code{AssayLink} instances with names
\code{names}.}
}
\description{
Links between assays within a \link{Features} object are handled by an
\code{AssayLinks} object. It is composed by a list of \code{AssayLink}
instances.
}
\details{
The constructor are respectively \code{AssayLinks(..., .Names = NULL)}
and \code{AssayLink(name, from = NULL, fcol = NULL, hits = NULL)}. See
examples below.
}
\examples{

al1 <- AssayLink(name = "assay1")
al1

AssayLinks(al1)

AssayLinks(names = c("Assay1", "Assay2"))
}
