% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayLinks.R
\name{AssayLinks}
\alias{AssayLinks}
\alias{AssayLink}
\alias{AssayLink-class}
\alias{AssayLinks-class}
\alias{class:AssayLinks}
\alias{class:AssayLink}
\alias{show,AssayLink-method}
\alias{[,AssayLink,character-method}
\alias{[,AssayLinks,character-method}
\alias{assayLink}
\alias{assayLinks}
\alias{[,AssayLink,character,ANY,ANY-method}
\alias{[,AssayLinks,list,ANY,ANY-method}
\alias{createAssayLink}
\alias{createAssayLinkOneToOne}
\title{Links between Assays}
\usage{
\S4method{show}{AssayLink}(object)

AssayLink(name, from = NA_character_, fcol = NA_character_, hits = Hits())

AssayLinks(..., names = NULL)

assayLink(x, i)

assayLinks(x, i)

\S4method{[}{AssayLink,character,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{AssayLinks,list,ANY,ANY}(x, i, j, ..., drop = TRUE)

createAssayLink(object, from, to, varFrom, varTo)

createAssayLinkOneToOne(object, from, to, varCommon)
}
\arguments{
\item{object}{An \code{AssayLink} object to show.}

\item{name}{A mandatory name of the assay(s).}

\item{from}{A character(1) or integer(1) indicating which assay to link from
in \code{object}}

\item{fcol}{The feature variable of the parent assay used to
generate the current assay (used in
\code{aggregateFeatures}). \code{NA_character_}, if not applicable.}

\item{hits}{An object of class \link[S4Vectors:Hits]{S4Vectors::Hits} matching the
features of two assays.}

\item{...}{A set of \code{AssayLink} objects or a list thereof.}

\item{names}{A \code{character()} of \code{AssayLink} names. If provided,
\code{...} are ignored, and \code{names} is used to create an
\code{AssayLinks} object with \code{AssayLink} instances with names
\code{names}.}

\item{x}{An instance of class \link{Features}.}

\item{i}{The index or name of the assay whose \code{AssayLink} and
parents \code{AssayLink} instances are to be returned. For \code{[}, the
feature names to filter on.}

\item{j}{ignored.}

\item{drop}{ignored.}

\item{to}{A character(1) or integer(1) indicating which assay to link to in
\code{object}}

\item{varFrom}{A character (1) indicating the feature variable to use to
match the \code{from} assay to the \code{to} assay.}

\item{varTo}{A character (1) indicating the feature variable to use to
match the \code{to} assay to the \code{from} assay. If missing, \code{varTo} is the
same as \code{varFrom}.}

\item{varCommon}{The name of the common feature variable in \code{from} and \code{to}.
If missing, row indexing will be used as matching variable.}
}
\value{
\code{assayLink} returns an instance of class \code{AssayLink}.

\code{assayLinks} returns an instance of class \code{AssayLinks}.
}
\description{
Links between assays within a \link{Features} object are handled by an
\code{AssayLinks} object. It is composed by a list of \code{AssayLink}
instances.
}
\section{Constructors}{


Object can be created with the \code{AssayLink()} and \code{AssayLinks()} constructors.
}

\section{Methods and functions}{

\itemize{
\item \code{assayLink(x, i)} accesses the AssayLink at position \code{i} or with
name \code{i} in the \link{Features} object \code{x}.
\item \code{parentAssayLinks(x, i, recursive = FALSE)} accesses the
parent(s) \code{AssayLinks} or assay with index or name \code{i}.
}
}

\section{Creating links between assays}{


\code{createAssayLink} and \code{createAssayLinkOneToOne} link two assays in a
Features object.
\itemize{
\item \code{createAssayLink} takes any two assays from the \code{Features} object and
creates a link given a matching feature variable in each assay's
\code{rowData}.
\item \code{createAssayLinkOneToOne} also links two assays from the \code{Features}
object, but the assays must have the same size and a one-to-one
relationship between the assays is created. A common feature variable can
be supplied, otherwise the rows of the assays are assumed to be matched.
}
}

\examples{

al1 <- AssayLink(name = "assay1")
al1

AssayLinks(al1)

al2 <- AssayLinks(names = c("Assay1", "Assay2"))
al2
}
