% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayLinks.R
\name{AssayLinks}
\alias{AssayLinks}
\alias{AssayLink}
\alias{AssayLink-class}
\alias{AssayLinks-class}
\alias{class:AssayLinks}
\alias{class:AssayLink}
\alias{show,AssayLink-method}
\alias{[,AssayLink,character-method}
\alias{[,AssayLinks,character-method}
\alias{assayLink}
\alias{assayLinks}
\alias{[,AssayLink,character,ANY,ANY-method}
\alias{[,AssayLinks,list,ANY,ANY-method}
\title{Links between Assays

Links between assays within a \link{Features} object are handled by an
\code{AssayLinks} object. It is composed by a list of \code{AssayLink}
instances.}
\usage{
\S4method{show}{AssayLink}(object)

AssayLink(name, from = NA_character_, fcol = NA_character_, hits = Hits())

AssayLinks(..., names = NULL)

assayLink(x, i)

assayLinks(x, i)

\S4method{[}{AssayLink,character,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{AssayLinks,list,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{An `AssayLink` object to show.}

\item{name}{A mandatory name of the assay(s).}

\item{from}{The name of the parent assay, or \code{NA_character_}, if
not applicable.}

\item{fcol}{The feature variable of the parent assay used to
generate the current assay (used in
\code{aggregateFeatures}). \code{NA_character_}, if not applicable.}

\item{hits}{An object of class \link[S4Vectors:Hits]{S4Vectors::Hits} matching the
features of two assays.}

\item{...}{A set of \code{AssayLink} objects or a list thereof.}

\item{names}{A \code{character()} of \code{AssayLink} names. If provided,
\code{...} are ignored, and \code{names} is used to create an
\code{AssayLinks} object with \code{AssayLink} instances with names
\code{names}.}

\item{x}{An instance of class \link{Features}.}

\item{i}{The index or name of the assay whose \code{AssayLink} and
parents \code{AssayLink} instances are to be returned. For \code{[}, the
feature names to filter on.}

\item{j}{ignored.}

\item{drop}{ignored.}
}
\value{
`assayLink` returns an instance of class `AssayLink`.

\code{assayLinks} returns an instance of class \code{AssayLinks}.
}
\description{
Links between Assays

Links between assays within a \link{Features} object are handled by an
\code{AssayLinks} object. It is composed by a list of \code{AssayLink}
instances.
}
\section{Constructors}{


Object can be created with the \code{AssayLink()} and \code{AssayLinks()} constructors.
}

\section{Methods and functions}{

\itemize{
\item \code{assayLink(x, i)} accesses the AssayLink at position \code{i} or with
name \code{i} in the \link{Features} object \code{x}.
\item \code{parentAssayLinks(x, i, recursive = FALSE)} accesses the
parent(s) \code{AssayLinks} or assay with index or name \code{i}.
}
}

\examples{

al1 <- AssayLink(name = "assay1")
al1

AssayLinks(al1)

al2 <- AssayLinks(names = c("Assay1", "Assay2"))
al2
}
