% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssayLinks.R
\name{AssayLinks}
\alias{AssayLinks}
\alias{EmptyAssayLinks}
\title{Links between Assays}
\usage{
EmptyAssayLinks()

AssayLinks(name, from = NULL, fcol = NULL)
}
\arguments{
\item{name}{The name of the assay(s).}

\item{from}{The name of the parent assay, or \code{NA_character_}, if
not applicable.}

\item{fcol}{The feature variable of the parent assay used to
generate the current assay (used in
\code{combineFeatures}). \code{NA_character_}, if not applicable.}
}
\description{
Links between assays within a \link{Features} object are handled by the
\code{links} slot. It is composed of a \code{DataFrame} with three columns,
documented below. \code{AssayLinks} can be created with the
\code{AssayLinks()} and \code{EmptyAssayLinks()} functions.
}
