% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feat1}
\alias{feat1}
\title{Test PSM-level expression data}
\format{An object of class \code{Features} of length 1.}
\usage{
feat1
}
\description{
A small test \code{Features} object for testing and demonstration.
}
\keyword{datasets}
