% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureList-class.R
\docType{class}
\name{FeatureList}
\alias{FeatureList}
\alias{FeatureList-class}
\alias{class:FeatureList}
\title{Contains a list of \code{FetureSet} instances}
\usage{
FeatureList(...)
}
\arguments{
\item{...}{Individual `FeatureSet` instances or a list thereof.}
}
\description{
This data structure as for sole purpose to store a list of
\link{FeatureSet} instances. All \code{FeatureSet} instances in a
\code{FeatureList} must however have the same number of columns.
Objects of the class can be constructed with the \code{FeatureList()}
constructor. The \code{[[} and \code{[} operators work as expected.
}
\details{
See \link{FeatureSet} for an example.
}
\author{
Laurent Gatto
}
