% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features-class.R, R/Features-constructors.R
\docType{methods}
\name{Features}
\alias{Features}
\alias{Features-class}
\alias{class:Features}
\alias{length,Features-method}
\alias{names,Features-method}
\alias{names<-,Features,character-method}
\alias{dims,Features-method}
\alias{dim,Features-method}
\alias{isEmpty,Features-method}
\alias{metadata,Features-method}
\alias{metadata<-,Features,list-method}
\alias{colData,Features-method}
\alias{colData<-,Features,DataFrame-method}
\alias{assays,Features-method}
\alias{assay,Features,numeric-method}
\alias{assay,Features,character-method}
\alias{assay,Features,missing-method}
\alias{featureData}
\alias{featureData,Features,ANY-method}
\alias{featureData,Features,numeric-method}
\alias{featureData,Features,character-method}
\alias{featureNames}
\alias{featureNames,Features,ANY-method}
\alias{featureNames,Features,numeric-method}
\alias{featureNames,Features,character-method}
\alias{featureVariables}
\alias{featureVariables,Features,ANY-method}
\alias{featureVariables,Features,numeric-method}
\alias{featureVariables,Features,character-method}
\alias{sampleNames,Features-method}
\alias{sampleNames<-,Features,character-method}
\title{Quantitative MS Features}
\usage{
\S4method{length}{Features}(x)

\S4method{names}{Features}(x)

\S4method{names}{Features,character}(x) <- value

\S4method{dims}{Features}(x)

\S4method{dim}{Features}(x)

\S4method{isEmpty}{Features}(x)

\S4method{metadata}{Features}(x, ...)

\S4method{metadata}{Features,list}(x) <- value

\S4method{colData}{Features}(x)

\S4method{colData}{Features,DataFrame}(x) <- value

\S4method{assays}{Features}(x, ..., withDimnames = TRUE)

\S4method{assay}{Features,numeric}(x, i, ...)

\S4method{assay}{Features,character}(x, i, ...)

\S4method{assay}{Features,missing}(x, i, ...)

featureData(x, i, ...)

\S4method{featureData}{Features,ANY}(x, i, ...)

\S4method{featureData}{Features,numeric}(x, i, ...)

\S4method{featureData}{Features,character}(x, i, ...)

featureNames(x, i, ...)

\S4method{featureNames}{Features,ANY}(x, i, ...)

\S4method{featureNames}{Features,numeric}(x, i, ...)

\S4method{featureNames}{Features,character}(x, i, ...)

featureVariables(x, i, ...)

\S4method{featureVariables}{Features,ANY}(x, i, ...)

\S4method{featureVariables}{Features,numeric}(x, i, ...)

\S4method{featureVariables}{Features,character}(x, i, ...)

\S4method{sampleNames}{Features}(object)

\S4method{sampleNames}{Features,character}(object) <- value

Features(assays = SimpleList(), featureData = SimpleList(),
  colData = DataFrame(), metadata = list())
}
\arguments{
\item{assays}{A `SimpleList` containing the object's assays (typically
matrices).}

\item{featureData}{A `SimpleList` containing the object's feature metadata
as `DataFrame` instances.}

\item{colData}{A `DataFrame` with column (sample annotations).}

\item{metadata}{A `list()` with arbitrary object annotations.}

\item{`x`:}{an instance of class \code{Features}.}

\item{`object`:}{an instance of class \code{Features}.}

\item{`i`:}{a \code{character(1)} or \code{numeric(1)} defining which assay of feature
data to access.}

\item{`value`:}{the replacement value.}
}
\value{
A new instance of class `Features`.
}
\description{
Conceptually, a \code{Features} object holds a set of \code{matrix} (or
\code{array}) elements containing quantitative data. The number of
columns (samples) are always the same across the matrices, but the
number of rows (features) can vary. Each one of these matrices has
a set of feature annotation (encoded as \code{DataFrame} objects), that
have the same number of rows as the assay matrix their are
associated to, and an arbitrary number of columns (feature
variables). In addition, a \code{Features} object also uses a single
\code{DataFrame} to annotate the samples (columns) represented in all
the matrices. The largest assay matrix (the one with the highest
number of features) is considered the main assay, from which the
other ones are derived by aggregating/combining several rows into
a single one.
}
\details{
A typical use case for such \code{Features} object if to represent
quantitative proteomics or metabolomics data, where different
assays represent quantitation data at the PSM (the main assay),
peptide and protein level, and where peptide values are computed
from the PSM data, and the protein-level data is calculated based
on the peptide-level values.

The recommended way to create \code{Features} objects is the use the
\code{readFeatures()} function, that creates an instance from tabular data. The
\code{Features} constructor can be used to create objects from their bare parts.
It iss the user's responsability to make sure that these match the class
validity requirements.

A \code{Features} instance must comply with the following requirements:
\itemize{
\item Assays are matrix-like objects that all have the same number of
columns. They are passed to the constructor as an \code{SimpleList} object
names \code{assays}, and can be accessed with the \code{assays} accessor.
\item Each assay is documented by feature variable \code{DataFrame}. Both have the
same number of rows. These feature metadata \code{DataFrame}s are passed to the
constructor as an \code{SimpleList} object names \code{featureData} and can be
accessed with the \code{featureData} accessor. The number of rows and the row
names of each of these \code{DataFrame} instances must match the number of rows
and rows names of the corresponding the assay. These \code{DataFrame} instance
are mandatory but can have no columns.
\item The samples/columns of the assays are documented in a \code{DataFrame} named
\code{colData} and can be accessed with the \code{colData} accessor. The number of
rows and the row names of this \code{DataFrame} must match the number of
columns and column names of the assays. This \code{DataFrame} instance is
mandatory but can have no columns.
\item A \code{list} protiding the global metadata annotating the object as a whole,
named \code{metadata}. It can be accessed/replaced with the \code{metadata}
accessor/replacement method.
}
}
\examples{

## Creating a Features object manually

m1 <- matrix(1:40, ncol = 4)
m2 <- matrix(1:16, ncol = 4)
colnames(m1) <- colnames(m2) <- paste0("S", 1:4)
rownames(m1) <- letters[1:10]
rownames(m2) <- letters[23:26]

df1 <- DataFrame(Fa = 1:10, Fb = letters[1:10],
                 row.names = letters[1:10])
df2 <- DataFrame(row.names = letters[23:26])

ft1 <- Features(assays = SimpleList(assay1 = m1, assay2 = m2),
                featureData = SimpleList(assay1 = df1, assay2 = df2),
                colData = DataFrame(row.names = colnames(m1)),
                metadata = list(paste("Generated on", Sys.Date())))
ft1

## Creating a Features object from a data.frame
data(hlpsms)
ft2 <- readFeatures(hlpsms, ecol = 1:10)
ft2

## The assay isn't named yet, so let's name it 'psms', to clarify that the
## data are PSM-level quantitations.
names(ft2)
names(ft2) <- "psms"
ft2
}
\author{
Laurent Gatto
}
