% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBy-methods.R
\name{subsetByFeature,Features,character-method}
\alias{subsetByFeature,Features,character-method}
\alias{subsetByFeature}
\title{Subset by feature name}
\usage{
\S4method{subsetByFeature}{Features,character}(x, y, ...)
}
\arguments{
\item{x}{An instance of class [Features].}

\item{y}{A `character` of feature names present in an assay in `x`.}

\item{...}{Additional parameters. Ignored.}
}
\value{
An new instance of class [Features] containing relevant
    assays and features.
}
\description{
This function will find the assays and features that match
directly (by name) or indirectly (through aggregation) the feature
name.

The `subsetByFeature` function will first identify the assay that
contains the feature(s) `i` and filter the rows matching these
feature names exactly. It will then find, in the other assays, the
features that produces `i` through aggregation with the
`aggregateFeatures` function.

See [Features] for an example.
}
\author{
Laurent Gatto
}
