% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSet-links.R
\docType{class}
\name{FeatureSetLink-class}
\alias{FeatureSetLink-class}
\alias{get_featureSet_ids}
\alias{set_featureSet_ids}
\alias{get_next_featureSet_id}
\title{Links between FeatureSets}
\usage{
get_featureSet_ids(object)

set_featureSet_ids(object)

get_next_featureSet_id(object)
}
\arguments{
\item{object}{A Features object.}
}
\value{
The \link{FeatureSet}s \code{id} slots.

A Features object with FeatureSet element ids set to
\code{1:length(object)}.

The next `id` in the [Features], as an `integer(1)`.
}
\description{
This class handles the slots to link different FeatureSets within
a Features object. It isn't meant to be used directly by users.
}
\section{Slots}{

\describe{
\item{\code{id}}{An \code{integer(1)} holding the set's identifier. For
internal use at the \link{Features}-level only.}

\item{\code{from}}{An \code{integer(1)} refering to the identifier from which
the current object originates. For internal use at the
\link{Features}-level only.}

\item{\code{fcol}}{A \code{character(1)} storing the feature variables that was
used to group the object's parent features.}
}}

