% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/missing-data.R
\name{zeroIsNA}
\alias{zeroIsNA}
\alias{missing-data}
\title{Managing missing data}
\usage{
zeroIsNA(object, ...)
}
\arguments{
\item{object}{An object of class \code{Features} or \code{SummarizedExperiment}.}
}
\value{
zeroIsNA zeroIsNA,SummarizedExperiment zeroIsNA,Features
}
\description{
This manual page describes the handling of missing values in
\code{Features} objects. The current functions are available:
\itemize{
\item \code{zeroIsNA(object)}: this function replaces all 0 in \code{object} by
\code{NA}. This is often necessary when third-party software assume
that features that weren't quantified should be assigned an
intensity of 0.
}
}
