% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures-aggregation.R
\name{aggregateSamples}
\alias{aggregateSamples}
\alias{aggregateSamples,QFeatures-method}
\alias{aggregateSamples,SummarizedExperiment-method}
\title{Aggregate assays' quantitative samples}
\usage{
\S4method{aggregateSamples}{SummarizedExperiment}(
  object,
  scol,
  fun = rowRobustSummary,
  moreFun = list(aggcounts = rowCounts),
  ...
)

\S4method{aggregateSamples}{QFeatures}(
  object,
  i,
  scol,
  name = "newAssay",
  fun = rowRobustSummary,
  moreFun = list(aggcounts = rowCounts),
  ...
)
}
\arguments{
\item{object}{An instance of class \link{QFeatures} or \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment}.}

\item{scol}{A \code{character()} naming the \code{colData} variable defining
how to group samples.}

\item{fun}{A function used to aggregate the samples. The function should
apply its summarization row-wise.}

\item{moreFun}{A named \code{list()} of additional aggregation functions
used to create additional assays. The function should apply its
summarization row-wise.}

\item{...}{Additional parameters passed to \code{fun}.}

\item{i}{A \code{character(1)} naming the assay that will be aggregated.}

\item{name}{A \code{character(1)} naming the new assay.}
}
\value{
A \code{QFeatures} object with additional assays or a
\code{SummarizedExperiment} object.
}
\description{
This function aggregates the quantitative samples of one or
multiple assays, applying a summarisation function (\code{fun}) to
sets of samples defined by a \code{colData} variable.
}
