% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures-aggregation.R
\name{aggregateSamples}
\alias{aggregateSamples}
\alias{aggregateSamples,QFeatures-method}
\alias{aggregateSamples,SummarizedExperiment-method}
\title{Aggregate sets' quantitative samples}
\usage{
\S4method{aggregateSamples}{SummarizedExperiment}(object, scol, fun = rowRobustSummary, ...)

\S4method{aggregateSamples}{QFeatures}(
  object,
  i,
  scol,
  name = "newAssay",
  fun = rowRobustSummary,
  ...
)
}
\arguments{
\item{object}{An instance of class \link{QFeatures} or \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment}.}

\item{scol}{A \code{character()} naming the \code{colData} variable defining
how to group samples.
When multiple variables are supplied, their values are concatenated
with \verb{_} to build the grouping variable.}

\item{fun}{A function used to aggregate the samples. The function should
apply its summarization row-wise.}

\item{...}{Additional parameters passed to \code{fun}.}

\item{i}{A \code{character(1)} naming the set that will be aggregated.}

\item{name}{A \code{character(1)} naming the new set.}
}
\value{
A \code{QFeatures} object with an additional set or a
\code{SummarizedExperiment} object. The new set/\code{SummarizedExperiment}
contains three assays the first named \code{assay} containing the result
of the aggregation, the second named \code{aggsd} containing the standard
deviation associated with each aggregated data point and the third
named \code{aggcounts} containing the number of observation used for each
aggregated data point.
}
\description{
This function aggregates quantitative samples, applying a
summarisation function (\code{fun}) to sets of samples defined by a
\code{colData} variable.

For \code{QFeatures} objects, the aggregated set \code{colData} is added to
the global \code{colData} using \code{\link[=addAssay]{addAssay()}}. If an aggregated group name
already exists as a row name in the global \code{colData}, it is reused
only when shared \code{colData} variables do not contain conflicting
values, otherwise an error is thrown. The aggregation of samples should
take place after run joining.
}
