% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures-constructors.R
\name{readQFeaturesFromDIANN}
\alias{readQFeaturesFromDIANN}
\title{Read DIA-NN output as a QFeatures objects}
\usage{
readQFeaturesFromDIANN(
  colAnnotation,
  assayData,
  extractedData = NULL,
  ecol = "Ms1.Area",
  multiplexing = "none",
  ...
)
}
\arguments{
\item{colAnnotation}{A \code{data.frame} or any object that can be
coerced to a \code{data.frame}. \code{colAnnotation} is expected to
contains all the sample annotations. We require the table to
contain a column called \code{File.Name} that links to the
\code{File.Name} in the DIA-NN report table. If \code{multiplexing = "mTRAQ"}, we require a second column called \code{Label} that links
the label to the sample (the labels identified by DIA-NN can
be retrieved from \verb{Modified Sequence} column in the report
table).}

\item{assayData}{A \code{data.frame} or any object that can be coerced
to a data.frame that contains the data from the \code{Report.tsv}
file generated by DIA-NN.}

\item{extractedData}{A data.frame or any object that can be coerced
to a data.frame that contains the data from the \verb{*_ms1_extracted.tsv}
file generated by DIA-NN. This argument is optional and is
currently only applicable for mTRAQ multiplexed experiments
where DIA-NN was run using the \code{plexdia} module.}

\item{ecol}{A \code{character(1)} indicating which column in
\code{assayData} contains the quantitative information. Default is
\code{"Ms1.Area"}.}

\item{multiplexing}{A \code{character(1)} indicating the type of
multiplexing used in the experiment. Provide \code{"none"} if the
experiment is label-free (default). Alternative options are:
\code{"mTRAQ"}.}

\item{...}{Further arguments passed to \code{\link[=readQFeatures]{readQFeatures()}}.}
}
\value{
An instance of class \code{QFeatures}. The quantiative data of
each acquisition run is stored in a separate set as a
\code{SummarizedExperiment} object.
\itemize{
\item The \code{QFeatures} (see \code{\link[=QFeatures]{QFeatures()}}) class to read about how to
manipulate the resulting \code{QFeatures} object.
\item The \code{\link[=readQFeatures]{readQFeatures()}} function to import other quantitative
data.
}
}
\description{
This function takes the output tables from DIA-NN and converts them
into a \code{QFeatures} object.
}
\examples{

x <- read.delim(MsDataHub::benchmarkingDIA.tsv())
## fix file names
x[[1]] <- sub("^.+raw-data\\\\\\\\", "", x[[1]])
cd <- data.frame(File.Name = unique(x[[1]]))
readQFeaturesFromDIANN(colAnnotation = cd, assayData = x)
}
\author{
Laurent Gatto, Christophe Vanderaa
}
